move Stopwatch into Logging
