+++ src/lenskit/logging/__init__.py
@@ -12,6 +12,7 @@ from ._console import console
 from ._proxy import get_logger
 from .config import LoggingConfig, basic_logging, notebook_logging
 from .progress import Progress, item_progress, set_progress_impl
+from .stopwatch import Stopwatch
 from .tasks import Task
 from .tracing import trace
 
@@ -26,4 +27,5 @@ __all__ = [
     "get_logger",
     "trace",
     "console",
+    "Stopwatch",
 ]
